Split metadata mapper in Attribute and EntityType mappers
